requirements.txt
streamlit
pandas
pandasai
pillow
pyyaml
google-generativeai
seaborn